Latest ffmpeg from XBMC:

+++ xbmc/cores/dvdplayer/Codecs/ffmpeg/build_xbmc_win32.sh
@@ -1,10 +1,12 @@
 #!/bin/bash 
+if [ -d .libs ]
+then
 rm -r .libs
+fi
+
 make distclean
 
-./configure \
---extra-cflags="-fno-common -I../libfaad2/include -Iinclude/dxva2" \
---extra-ldflags="-L../../../../../system/players/dvdplayer" \
+OPTIONS="
 --enable-shared \
 --enable-memalign-hack \
 --enable-gpl \
@@ -17,10 +19,19 @@ make distclean
 --disable-muxers \
 --disable-encoders \
 --disable-debug \
+--disable-ffplay \
+--disable-ffserver \
+--disable-ffmpeg \
+--disable-ffprobe \
+--disable-devices \
 --enable-muxer=spdif \
 --enable-muxer=adts \
 --enable-encoder=ac3 \
---enable-encoder=aac &&
+--enable-encoder=aac \
+--enable-runtime-cpudetect \
+--disable-debug"
+
+./configure --extra-cflags="-fno-common -I../libfaad2/include -Iinclude/dxva2" --extra-ldflags="-L../../../../../system/players/dvdplayer" ${OPTIONS} &&
  
 make -j3 && 
 mkdir .libs &&
